main

+++ SistemaTrasporti/src/main/java/Entities/RivenditoreAutorizzato.java
@@ -1,33 +1,30 @@
 package Entities;
 
 import javax.persistence.Column;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 
-
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
 public class RivenditoreAutorizzato {
-	
+
 	@Id
 	@GeneratedValue
-	private int id;
-	
+	private Long id;
+
 	@Column(name = "nome_rivenditore")
 	private String nome;
-	
+
 	private String indirizzo;
-	
-	public RivenditoreAutorizzato(String nome ,String indirizzo) {
+
+	public RivenditoreAutorizzato(String nome, String indirizzo) {
 		this.nome = nome;
 		this.indirizzo = indirizzo;
 	}
@@ -36,6 +33,5 @@ public class RivenditoreAutorizzato {
 	public String toString() {
 		return "RivenditoreAutorizzato [nome=" + nome + ", indirizzo=" + indirizzo + "]";
 	}
-	
-	
+
 }
