+++ SistemaTrasporti/src/main/java/Entities/Tessera.java
@@ -2,7 +2,6 @@ package Entities;
 
 import java.time.LocalDate;
 
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
@@ -11,27 +10,26 @@ import javax.persistence.OneToOne;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
 public class Tessera {
-	
+
 	private LocalDate dataScadenza;
-	
+
 	@Id
-	@GeneratedValue 
-	private int id;
-	
+	@GeneratedValue
+	private Long id;
+
 	private Long numeroTessera;
-	
+
 	public Tessera(LocalDate _dataScadenza, Long numeroTessera) {
 		this.dataScadenza = _dataScadenza;
 		this.numeroTessera = numeroTessera;
 	}
-	
+
 	@OneToOne(mappedBy = "tessera")
 	private Utente utente;
 }
